pyproject.toml
[tool.poetry]
name = "gql"
version = "0.1.0"
description = "Python GraphQL Client Library"
authors = ["Eran Kampf <eran@ekampf.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
click = "^7.0"
graphql-core-next = "~=1.0.1"
requests = "^2.21"
jinja2 = "^2.10"
inflection = "^0.3.1"
dataclasses-json = "^0.2.0"

[tool.poetry.dev-dependencies]
pylint = "^2.2.2"
pytest = "^4.0.1"
pytest-describe = "*"
pytest-expecter = "*"
pytest-random = "*"
pytest-cov = "*"
freezegun = "*"
verchew = "^1.4"
pre_commit = "~=1.12.0"
ipython = "^7.0"

[tool.poetry.scripts]
my-script = 'gql:cli'

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
